requirements.txt
gradio
onnxruntime
torch
torchvision
transformers
open_clip_torch
numpy
scipy
scikit-image
Pillow
redis
plotly
python-dotenv